requirements.txt
python-dotenv
openai
autogen-agentchat
autogen-ext
traceloop-sdk
